(C++) clang-format plus remove a warning.

+++ C++/Multiplayer/Step 5 - Weapons/client/components/AnimatedSprite.hpp
@@ -24,13 +24,11 @@ namespace components
         {
         }
 
-        void updateElapsedTime(std::chrono::microseconds howMuch) {
-            m_elapsedTime += howMuch;
-        }
+        void updateElapsedTime(std::chrono::microseconds howMuch) { m_elapsedTime += howMuch; }
         auto get() { return m_sprites[m_currentSprite]; }
         auto getCurrentSpriteTime() { return m_spriteTime[m_currentSprite]; };
         auto getElapsedTime() { return m_elapsedTime; }
-        void incrementSprite() { m_currentSprite = (m_currentSprite + 1) % m_sprites.size(); }
+        void incrementSprite() { m_currentSprite = (m_currentSprite + static_cast<std::uint8_t>(1)) % m_sprites.size(); }
 
       private:
         std::vector<std::shared_ptr<sf::Sprite>> m_sprites;
